requirements.txt
tenacity
termcolor
openai
python-dotenv
bs4
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
sqlalchemy==1.3.24
sqlalchemy_utils
jsonpath_ng
markupsafe==1.1.1
pygithub
itsdangerous==0.24
werkzeug
